add routing folder and declutter server
