+++ views/partials/header.ejs
@@ -11,7 +11,7 @@
           </li>
           <% if (loggedIn) { %>
             <li class="nav-item">
-                <a class="nav-link" href="/profile">Profile</a>
+                <a class="nav-link" href="/user">Profile</a>
               </li>
         <% }%>
         </ul>
